- TimeZoneInfo::ConvertTime 함수 구현 및 result 값 문제 있는 것 수정 - DateTime::GetKind 반환 값이 제대로 되지 않은 문제 수정. 63, 64비트 부분을 kind로 사용하므로 m_ticks 타입을 unsigned로 바꿈 - WindowsFile::SetLastTimeUtc 함수 구현

+++ Source/TGON/String/Encoding.h
@@ -65,7 +65,7 @@ inline int32_t Encoding<_FromEncodingType>::ConvertTo(const std::basic_string_vi
 
     int32_t encodedStrBytes = ustr.extract(0, ustr.length(), nullptr, _ToEncodingType::EncodingName);
     int32_t destStrBufferBytes = destStrBufferLen * sizeof(_DestCharType);
-    if (encodedStrBytes + sizeof(_DestCharType) > destStrBufferBytes)
+    if (encodedStrBytes + static_cast<int32_t>(sizeof(_DestCharType)) > destStrBufferBytes)
     {
         return -1;
     }
